refactor: sign up success modal and email verification modal
