+++ src/context/modals.tsx
@@ -1,11 +1,15 @@
+import MainButton from 'components/Button/MainButton'
+import OutlineTextField from 'components/Input/TextField/Outline'
 import Modal from 'components/Modal'
 import ForgotPasswordModal from 'components/Modal/ForgotPasswordModal'
 import MigrateWalletModal from 'components/Modal/MigrateWalletModal'
 import SignInModal from 'components/Modal/SignInModal'
 import SignUpModal from 'components/Modal/SignUpModal'
 import SignUpSuccessModal from 'components/Modal/SignUpSuccessModal'
-import { Dispatch, SetStateAction, createContext, useState } from 'react'
-
+import { Dispatch, SetStateAction, createContext, useContext, useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import { validateEmail } from 'src/utils'
+import { Context } from '.'
 export const ModalContext = createContext<{
   signUpSuccessOpen: boolean
   forgotPasswordOpen: boolean
@@ -17,6 +21,7 @@ export const ModalContext = createContext<{
   setSignUpOpen: Dispatch<SetStateAction<boolean>>
   setSignInOpen: Dispatch<SetStateAction<boolean>>
   setMigrateWalletOpen: Dispatch<SetStateAction<boolean>>
+  showEmailVerification: (email: string) => void
 }>({
   signUpSuccessOpen: false,
   forgotPasswordOpen: false,
@@ -28,6 +33,7 @@ export const ModalContext = createContext<{
   setSignUpOpen: () => {},
   setSignInOpen: () => {},
   setMigrateWalletOpen: () => {},
+  showEmailVerification: () => {},
 })
 function ModalProvider({ children }) {
   const [signUpSuccessOpen, setSignUpSuccessOpen] = useState(false)
@@ -35,6 +41,75 @@ function ModalProvider({ children }) {
   const [signUpOpen, setSignUpOpen] = useState(false)
   const [signInOpen, setSignInOpen] = useState(false)
   const [migrateWalletOpen, setMigrateWalletOpen] = useState(false)
+  const { account, updateProfile } = useContext(Context)
+  const [errorMsg, setErrorMsg] = useState('')
+  const [emailErrorMsg, setEmailErrorMsg] = useState('')
+  const [name, setName] = useState('')
+  const [email, setEmail] = useState('')
+  const [loading, setLoading] = useState(false)
+  const [open, setOpen] = useState(true)
+  const [emailNeedVerify, setEmailNeedVerify] = useState('')
+  const { t } = useTranslation()
+
+  useEffect(() => {
+    setErrorMsg('')
+  }, [name])
+
+  useEffect(() => {
+    setEmailErrorMsg('')
+  }, [email])
+
+  useEffect(() => {
+    if (!signUpSuccessOpen) {
+      setTimeout(() => setEmailNeedVerify(''), 1000)
+    }
+  }, [signUpSuccessOpen])
+
+  const setUName = async () => {
+    try {
+      setLoading(true)
+      await updateProfile({
+        nickname: name,
+      })
+      setLoading(false)
+    } catch (error) {
+      setErrorMsg(t('Name already taken'))
+      setLoading(false)
+    }
+  }
+  const setUNameAndEmail = async () => {
+    try {
+      if (!validateEmail(email)) {
+        setEmailErrorMsg('Invalid email address')
+        return
+      }
+      setLoading(true)
+      await updateProfile({
+        nickname: name,
+        email,
+      })
+      setLoading(false)
+      setOpen(false)
+      showEmailVerification(email)
+    } catch (error) {
+      setLoading(false)
+      if (error.message?.includes('authorizer_users_nickname_key')) {
+        setErrorMsg(t('Name already taken'))
+        return
+      }
+      if (error.message?.includes('email address already exists')) {
+        setEmailErrorMsg(t('Email has been registered'))
+        return
+      }
+      setErrorMsg(t('Something went wrong'))
+    }
+  }
+
+  const showEmailVerification = (email: string) => {
+    setEmailNeedVerify(email)
+    setSignUpSuccessOpen(true)
+  }
+
   return (
     <ModalContext.Provider
       value={{
@@ -48,6 +123,7 @@ function ModalProvider({ children }) {
         setSignUpOpen,
         setSignInOpen,
         setMigrateWalletOpen,
+        showEmailVerification,
       }}>
       <Modal
         preventClickOutsideToClose
@@ -72,7 +148,7 @@ function ModalProvider({ children }) {
               setSignUpOpen(false)
             }}
             setSignUpOpen={setSignUpOpen}
-            setSignUpSuccessOpen={setSignUpSuccessOpen}
+            showEmailVerification={showEmailVerification}
           />
         </div>
       </Modal>
@@ -81,9 +157,96 @@ function ModalProvider({ children }) {
         <ForgotPasswordModal onClose={() => setForgotPasswordOpen(false)} />
       </Modal>
       <Modal open={signUpSuccessOpen} setOpen={setSignUpSuccessOpen}>
-        <SignUpSuccessModal setSignUpOpen={setSignUpOpen} onClose={() => setSignUpSuccessOpen(false)} />
+        <SignUpSuccessModal
+          setSignUpOpen={setSignUpOpen}
+          onClose={() => setSignUpSuccessOpen(false)}
+          email={emailNeedVerify}
+        />
       </Modal>
       {migrateWalletOpen && <MigrateWalletModal open={migrateWalletOpen} setOpen={setMigrateWalletOpen} />}
+
+      {account ? (
+        validateEmail(account?.email) ? (
+          account?.name ? (
+            <></>
+          ) : (
+            <Modal open={true} setOpen={setOpen}>
+              <div className='p-6 flex flex-col w-[322px]'>
+                <div className='text-xl font-bold leading-6 text-center'>{t('Set a username')}</div>
+                <div className='mt-6'>
+                  <OutlineTextField label={t('Username')} errorMsg={errorMsg} value={name} onChange={setName} />
+                  <OutlineTextField label={t('Email')} value={account?.email} disabled={true} />
+                </div>
+                <p className='text-[10px] -mt-4'>
+                  {t('This email will also be used to receive updates of new chapter when you subscribe a manga.')}
+                </p>
+                <div className='mt-4'>
+                  <MainButton className='w-full' disabled={!name} loading={loading} onClick={setUName}>
+                    {t('Continue')}
+                  </MainButton>
+                </div>
+                <p className='text-xs mt-1 text-center'>
+                  {t('Or')}{' '}
+                  <a
+                    className='text-second-color'
+                    onClick={() => {
+                      setOpen(false)
+                      setSignInOpen(true)
+                    }}>
+                    {t('sign in')}
+                  </a>{' '}
+                  {t('with another account')}
+                </p>
+              </div>
+            </Modal>
+          )
+        ) : (
+          <Modal open={open} setOpen={setOpen}>
+            <div className='p-6 flex flex-col w-[322px]'>
+              <div className='text-xl font-bold leading-6 text-center'>{t('Verify your email')}</div>
+              <p className='text-xs leading-4 mt-1 text-center'>
+                {t(
+                  'An active email is required when sign in to Punkga, verify it only once and enjoy all of our great mangas.'
+                )}
+              </p>
+              <div className='mt-6'>
+                <OutlineTextField
+                  label={t('Email')}
+                  value={email}
+                  onChange={setEmail}
+                  errorMsg={emailErrorMsg}
+                  placeholder={t('Enter your email')}
+                />
+                <OutlineTextField
+                  label={t('Username')}
+                  errorMsg={errorMsg}
+                  value={name}
+                  onChange={setName}
+                  placeholder={t('Enter username')}
+                />
+              </div>
+              <MainButton className='w-full mt-6' disabled={!name} loading={loading} onClick={setUNameAndEmail}>
+                {t('Continue')}
+              </MainButton>
+              <p className='text-xs mt-1 text-center'>
+                {t('Or')}{' '}
+                <a
+                  className='text-second-color '
+                  onClick={() => {
+                    setOpen(false)
+                    setSignInOpen(true)
+                  }}>
+                  {t('sign in')}
+                </a>{' '}
+                {t('with another account')}
+              </p>
+            </div>
+          </Modal>
+        )
+      ) : (
+        <></>
+      )}
+
       {children}
     </ModalContext.Provider>
   )
