+++ src/pages/_app.tsx
@@ -5,19 +5,15 @@ import { ChainProvider } from '@cosmos-kit/react'
 import '@interchain-ui/react/styles'
 import axios from 'axios'
 import { assets, chains } from 'chain-registry'
-import FilledButton from 'components/Button/FilledButton'
 import HeadComponent from 'components/Head'
-import OutlineTextField from 'components/Input/TextField/Outline'
-import Modal from 'components/Modal'
 import MaintainPage from 'components/pages/maintainPage'
 import moment from 'moment'
 import 'moment/locale/vi'
-import { appWithTranslation, useTranslation } from 'next-i18next'
+import { appWithTranslation } from 'next-i18next'
 import { AppProps } from 'next/app'
 import getConfig, { setConfig } from 'next/config'
 import { Plus_Jakarta_Sans, Work_Sans } from 'next/font/google'
 import localFont from 'next/font/local'
-import Image from 'next/image'
 import { useRouter } from 'next/router'
 import { useContext, useEffect, useState } from 'react'
 import { isMobile } from 'react-device-detect'
@@ -25,11 +21,10 @@ import { ToastContainer } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
 import 'slick-carousel/slick/slick-theme.css'
 import 'slick-carousel/slick/slick.css'
-import Mail from 'src/assets/images/Mail.svg'
 import ContextProvider, { Context } from 'src/context'
 import { wallets as c98Mobile } from 'src/services/c98MobileWallet'
 import 'src/styles/globals.scss'
-import { getGasPriceByChain, validateEmail } from 'src/utils'
+import { getGasPriceByChain } from 'src/utils'
 const pjs = Plus_Jakarta_Sans({ subsets: ['latin', 'vietnamese'] })
 const ws = Work_Sans({ subsets: ['latin', 'vietnamese'] })
 const orbitron = localFont({
@@ -182,22 +177,7 @@ function MyApp(props: AppProps) {
   )
 }
 const App = ({ Component, pageProps }: AppProps) => {
-  const { isSettingUp, account, updateProfile, resendVerifyEmail } = useContext(Context)
-  const [errorMsg, setErrorMsg] = useState('')
-  const [emailErrorMsg, setEmailErrorMsg] = useState('')
-  const [name, setName] = useState('')
-  const [email, setEmail] = useState('')
-  const [loading, setLoading] = useState(false)
-  const [open, setOpen] = useState(true)
-  const [openSuccessModal, setOpenSuccessModal] = useState(false)
-  const { t } = useTranslation()
-
-  useEffect(() => {
-    setErrorMsg('')
-  }, [name])
-  useEffect(() => {
-    setEmailErrorMsg('')
-  }, [email])
+  const { isSettingUp } = useContext(Context)
 
   useEffect(() => {
     const scriptElement = document.createElement('script')
@@ -214,177 +194,10 @@ const App = ({ Component, pageProps }: AppProps) => {
     document.body.append(noscriptElement)
   }, [])
 
-  useEffect(() => {
-    if (account && validateEmail(account?.email) && account?.name && !account?.verified) {
-      setOpenSuccessModal(true)
-    }
-  }, [account])
-
   if (isSettingUp) {
     return <></>
   }
 
-  const setUName = async () => {
-    try {
-      setLoading(true)
-      await updateProfile({
-        nickname: name,
-      })
-      setLoading(false)
-    } catch (error) {
-      setErrorMsg(t('Name already taken'))
-      setLoading(false)
-    }
-  }
-  const setUNameAndEmail = async () => {
-    try {
-      if (!validateEmail(email)) {
-        setEmailErrorMsg('Invalid email address')
-        return
-      }
-      setLoading(true)
-      await updateProfile({
-        nickname: name,
-        email,
-      })
-      setLoading(false)
-      setOpen(false)
-      setOpenSuccessModal(true)
-    } catch (error) {
-      setLoading(false)
-      if (error.message?.includes('authorizer_users_nickname_key')) {
-        setErrorMsg(t('Name already taken'))
-        return
-      }
-      if (error.message?.includes('email address already exists')) {
-        setEmailErrorMsg(t('Email has been registered'))
-        return
-      }
-      setErrorMsg(t('Something went wrong'))
-    }
-  }
-
-  return (
-    <>
-      {account ? (
-        validateEmail(account?.email) ? (
-          account?.name ? (
-            <></>
-          ) : (
-            <Modal open={open} setOpen={setOpen}>
-              <div className='p-6 flex flex-col w-[322px]'>
-                <div className='gap-2 flex flex-col'>
-                  <div className='text-xl font-semibold leading-6 text-center'>{t('Set a username')}</div>
-                  <OutlineTextField label={t('Username')} errorMsg={errorMsg} value={name} onChange={setName} />
-                  <OutlineTextField label={t('Email')} value={account.email} disabled={true} />
-                </div>
-                <p className='text-xs mt-[6px] text-center'>
-                  {t('This email will also be used to receive updates of new chapter when you subscribe a manga.')}
-                </p>
-                <div className='mt-3 mx-auto'>
-                  <FilledButton size='lg' disabled={!name} loading={loading} onClick={setUName}>
-                    {t('Continue')}
-                  </FilledButton>
-                </div>
-                <p className='text-xs mt-2 font-medium text-center'>
-                  {t('Or')}{' '}
-                  <a
-                    className='text-second-color font-bold'
-                    onClick={() => {
-                      setOpen(false)
-                      document.getElementById('open-sign-in-btn')?.click()
-                    }}>
-                    {t('sign in')}
-                  </a>{' '}
-                  {t('with another account')}
-                </p>
-              </div>
-            </Modal>
-          )
-        ) : (
-          <Modal open={open} setOpen={setOpen}>
-            <div className='p-6 flex flex-col w-[322px]'>
-              <div className='gap-3 flex flex-col'>
-                <div className='text-xl font-semibold leading-6 text-center'>{t('Verify your email')}</div>
-                <p className='text-[10px] leading-3 text-center'>
-                  {t(
-                    'An active email is required when sign in to Punkga, verify it only once and enjoy all of our great mangas.'
-                  )}
-                </p>
-                <OutlineTextField
-                  label={t('Email')}
-                  value={email}
-                  onChange={setEmail}
-                  errorMsg={emailErrorMsg}
-                  placeholder={t('Enter your email')}
-                />
-                <OutlineTextField
-                  label={t('Username')}
-                  errorMsg={errorMsg}
-                  value={name}
-                  onChange={setName}
-                  placeholder={t('Enter username')}
-                />
-              </div>
-              <div className='mt-3 mx-auto'>
-                <FilledButton size='lg' disabled={!name} loading={loading} onClick={setUNameAndEmail}>
-                  {t('Continue')}
-                </FilledButton>
-              </div>
-              <p className='text-xs mt-2 font-medium text-center'>
-                {t('Or')}{' '}
-                <a
-                  className='text-second-color font-bold'
-                  onClick={() => {
-                    setOpen(false)
-                    document.getElementById('open-sign-in-btn')?.click()
-                  }}>
-                  {t('sign in')}
-                </a>{' '}
-                {t('with another account')}
-              </p>
-            </div>
-          </Modal>
-        )
-      ) : (
-        <></>
-      )}
-      <Modal open={openSuccessModal} setOpen={setOpenSuccessModal}>
-        <div className={` py-6 px-[60px] flex flex-col gap-4 w-full max-w-[670px]`}>
-          <p className='text-center text-xl leading-6 font-bold'>{t('Email verification')}</p>
-          <Image src={Mail} alt='' className='mx-auto' />
-          <p className=' font-medium text-center w-full max-w-[500px] mx-auto'>
-            {t('A verification link has been sent to')}{' '}
-            <span className='text-second-color font-bold'>{account?.email || email}</span>.
-            {t('Please click on the link to verify your email account.')}
-          </p>
-          <div className='flex flex-col text-center text-xs leading-[14px]'>
-            <p className=' font-medium text-center w-full max-w-[500px] mx-auto'>
-              {t('Have not received any email')}?{' '}
-              <span
-                className='text-second-color font-bold cursor-pointer'
-                onClick={() => resendVerifyEmail(account?.email || email, 'update_email')}>
-                {t('Click here')}
-              </span>{' '}
-              {t('to resend verification link')}
-            </p>
-            <p className=' font-medium text-center w-full max-w-[500px] mx-auto'>
-              {t('Or')}{' '}
-              <span
-                className='text-second-color font-bold cursor-pointer'
-                onClick={() => {
-                  setOpenSuccessModal(false)
-                  document.getElementById('open-sign-in-btn')?.click()
-                }}>
-                {t('sign up')}
-              </span>{' '}
-              {t('with another email')}
-            </p>
-          </div>
-        </div>
-      </Modal>
-      <Component {...pageProps} />
-    </>
-  )
+  return <Component {...pageProps} />
 }
 export default appWithTranslation(MyApp)
